% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layer.R
\docType{class}
\name{Layer}
\alias{Layer}
\title{Class that encapsulates a layer used for dispalying slices of a 3D volume.}
\format{\code{\link{R6Class}} object.}
\usage{
Layer
}
\value{
Object of \code{\link{R6Class}} with methods for accessing and changing layer properties.
}
\description{
Class that encapsulates a layer used for dispalying slices of a 3D volume.
}
\section{Fields}{

\describe{
\item{\code{vol}}{an object of class \code{\linkS4class{NeuroVol}} supplying the image data}

\item{\code{color_map}}{the color map}

\item{\code{threshold}}{the low and high thresholds to determine transparency.}

\item{\code{irange}}{the intensity range of the low and high end of color scale.}

\item{\code{view}}{the orientation code (e.g. "LPI") describing the display axes.}

\item{\code{zero_col}}{the color assigned to values of 0.}

\item{\code{alpha}}{the degree of opacity for the layer (0 = transparent, 1 = opaque)}

\item{\code{view_space}}{the \code{linkS4class{NeuroSpace}} associaed with the view.}

\item{\code{value_range}}{the full numeric range of the underlying volume}

\item{\code{desc}}{a short \code{character} description of the layer}
}}

\keyword{datasets}
