% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrthoPlot.R
\name{ortho_plot}
\alias{ortho_plot}
\title{ortho_plot}
\usage{
ortho_plot(..., height = 300)
}
\arguments{
\item{...}{a variable length list of \code{\linkS4class{NeuroVol}} objects}
}
\description{
A shiny-based orthogonal viewer for a volumetric neuroimaging data
}
