% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layer.R
\docType{class}
\name{Overlay}
\alias{Overlay}
\title{Class that describes a stack of image Layers}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Overlay
}
\description{
Class that describes a stack of image Layers
}
\section{Fields}{

\describe{
\item{\code{layers}}{a \code{list} of \code{Layer} objects}

\item{\code{layer_names}}{}
}}

\keyword{datasets}
