% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layer.R
\docType{class}
\name{RenderedSliceStack}
\alias{RenderedSliceStack}
\title{Class representing a stack of rendered slices.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RenderedSliceStack
}
\description{
Class representing a stack of rendered slices.
}
\keyword{datasets}
