% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrthoPlot.R
\name{slice_plot}
\alias{slice_plot}
\title{slice_plot}
\usage{
slice_plot(overlay, height = 300)
}
\arguments{
\item{overlay}{a \code{Overlay} object}
}
\description{
A shiny-based slice-based viewer for a volumetric neuroimaging data
}
